+++ A6/src/Transform.h
@@ -3,10 +3,10 @@
 #include <glm/glm.hpp>
 
 struct Transform {
-	glm::vec4 loc = glm::vec4(0, 0, 0, 1);
-	glm::vec3 rot = glm::vec3(0, 0, 0);
-	glm::vec3 scale = glm::vec3(1, 1, 1);
+	glm::dvec4 loc = glm::dvec4(0, 0, 0, 1);
+	glm::dvec3 rot = glm::dvec3(0, 0, 0);
+	glm::dvec3 scale = glm::dvec3(1, 1, 1);
 	Transform() = default;
-	Transform(glm::vec4 _loc, glm::vec3 _rot, glm::vec3 _scale)
+	Transform(glm::dvec4 _loc, glm::dvec3 _rot, glm::dvec3 _scale)
 		: loc(_loc), rot(_rot), scale(_scale) {}
 };
